Split Python runtime from build packaging
