+++ gestaltd/internal/pluginpkg/python_provider.go
@@ -17,6 +17,7 @@ import (
 const (
 	pythonProjectFile   = "pyproject.toml"
 	pythonRuntimeModule = "gestalt._runtime"
+	pythonBuildModule   = "gestalt._build"
 )
 
 var ErrNoPythonProviderPackage = errors.New("no Python provider package found")
@@ -65,7 +66,7 @@ func BuildPythonProviderBinary(sourceDir, binaryPath, pluginName, target string)
 		return fmt.Errorf("detect Python release interpreter: %w", err)
 	}
 
-	cmd := exec.Command(interpreter, "-m", pythonRuntimeModule, "build", sourceDir, target, binaryPath, pluginName)
+	cmd := exec.Command(interpreter, "-m", pythonBuildModule, sourceDir, target, binaryPath, pluginName)
 	cmd.Dir = sourceDir
 	cmd.Env = append(os.Environ(), pythonBackendEnv()...)
 	cmd.Stdout = os.Stdout
