+++ sdk/python/gestalt/_runtime.py
@@ -4,9 +4,7 @@ import json
 import os
 import pathlib
 import signal
-import subprocess
 import sys
-import tempfile
 from concurrent import futures
 
 from ._plugin import ENV_WRITE_CATALOG, Plugin, Request
@@ -95,13 +93,6 @@ def main(argv: list[str] | None = None) -> int:
         if plugin_name is not None:
             plugin_name = str(plugin_name).strip() or None
         root = None
-    elif args[0] == "build":
-        if len(args) != 5:
-            _print_usage()
-            return 2
-        _, root, target, output_path, plugin_name = args
-        build_plugin_binary(root, target, output_path, plugin_name)
-        return 0
     else:
         if len(args) != 2:
             _print_usage()
@@ -119,56 +110,6 @@ def main(argv: list[str] | None = None) -> int:
 
     serve(plugin)
     return 0
-
-
-def build_plugin_binary(root: str, target: str, output_path: str, plugin_name: str) -> None:
-    root_path = pathlib.Path(root).resolve()
-    output = pathlib.Path(output_path).resolve()
-    module_name, _attr_name = _split_target(target)
-
-    output.parent.mkdir(parents=True, exist_ok=True)
-    with tempfile.TemporaryDirectory(prefix="gestalt-python-release-") as work_dir:
-        work_path = pathlib.Path(work_dir)
-        bundle_config_path = work_path / BUNDLED_CONFIG_NAME
-        bundle_config_path.write_text(
-            json.dumps(
-                {
-                    "target": target,
-                    "plugin_name": plugin_name,
-                }
-            ),
-            encoding="utf-8",
-        )
-
-        pyinstaller_name = output.name
-        if sys.platform == "win32" and pyinstaller_name.endswith(".exe"):
-            pyinstaller_name = pyinstaller_name[:-4]
-        separator = ";" if sys.platform == "win32" else ":"
-
-        command = [
-            sys.executable,
-            "-m",
-            "PyInstaller",
-            "--noconfirm",
-            "--clean",
-            "--onefile",
-            "--distpath",
-            str(output.parent),
-            "--workpath",
-            str(work_path / "build"),
-            "--specpath",
-            str(work_path / "spec"),
-            "--name",
-            pyinstaller_name,
-            "--hidden-import",
-            module_name,
-            "--paths",
-            str(root_path),
-            "--add-data",
-            f"{bundle_config_path}{separator}{BUNDLED_CONFIG_NAME}",
-            str(pathlib.Path(__file__).with_name("_pyinstaller.py")),
-        ]
-        subprocess.run(command, cwd=root_path, check=True)
 
 
 def _load_plugin(target: str, root: str | None = None) -> Plugin:
@@ -194,8 +135,7 @@ def _split_target(target: str) -> tuple[str, str]:
 
 def _print_usage() -> None:
     print(
-        "usage: python -m gestalt._runtime ROOT MODULE:ATTRIBUTE\n"
-        "   or: python -m gestalt._runtime build ROOT MODULE:ATTRIBUTE OUTPUT PLUGIN_NAME",
+        "usage: python -m gestalt._runtime ROOT MODULE:ATTRIBUTE",
         file=sys.stderr,
     )
